+++ backend/src/baserow/contrib/database/api/tables/views.py
@@ -1,4 +1,5 @@
 from django.db import transaction
+from baserow.api.sessions import get_untrusted_client_session_id
 from drf_spectacular.openapi import OpenApiParameter, OpenApiTypes
 from drf_spectacular.utils import extend_schema
 from rest_framework.permissions import IsAuthenticated
@@ -10,28 +11,13 @@ from baserow.api.decorators import validate_body, map_exceptions
 from baserow.api.errors import ERROR_USER_NOT_IN_GROUP
 from baserow.api.schemas import get_error_schema, CLIENT_SESSION_ID_SCHEMA_PARAMETER
 from baserow.api.trash.errors import ERROR_CANNOT_DELETE_ALREADY_DELETED_ITEM
-from baserow.contrib.database.api.fields.errors import (
-    ERROR_MAX_FIELD_COUNT_EXCEEDED,
-    ERROR_MAX_FIELD_NAME_LENGTH_EXCEEDED,
-    ERROR_RESERVED_BASEROW_FIELD_NAME,
-    ERROR_INVALID_BASEROW_FIELD_NAME,
-)
-from baserow.contrib.database.fields.exceptions import (
-    MaxFieldLimitExceeded,
-    MaxFieldNameLengthExceeded,
-    ReservedBaserowFieldNameException,
-    InvalidBaserowFieldName,
-)
+
 from baserow.contrib.database.handler import DatabaseHandler
 from baserow.contrib.database.table.exceptions import (
     TableDoesNotExist,
     TableNotInDatabase,
-    InvalidInitialTableData,
-    InitialTableDataLimitExceeded,
-    InitialTableDataDuplicateName,
 )
 from baserow.contrib.database.table.actions import (
-    CreateTableActionType,
     DeleteTableActionType,
     OrderTableActionType,
     UpdateTableActionType,
@@ -47,16 +33,12 @@ from baserow.api.jobs.errors import ERROR_MAX_JOB_COUNT_EXCEEDED
 from .errors import (
     ERROR_TABLE_DOES_NOT_EXIST,
     ERROR_TABLE_NOT_IN_DATABASE,
-    ERROR_INVALID_INITIAL_TABLE_DATA,
-    ERROR_INITIAL_TABLE_DATA_LIMIT_EXCEEDED,
-    ERROR_INITIAL_TABLE_DATA_HAS_DUPLICATE_NAMES,
 )
 from .serializers import (
     TableSerializer,
     TableCreateSerializer,
     TableUpdateSerializer,
     OrderTablesSerializer,
-    TableWithImportJobsSerializer,
 )
 
 
@@ -84,7 +66,7 @@ class TablesView(APIView):
             "endpoints."
         ),
         responses={
-            200: TableWithImportJobsSerializer(many=True),
+            200: TableSerializer(many=True),
             400: get_error_schema(["ERROR_USER_NOT_IN_GROUP"]),
             404: get_error_schema(["ERROR_APPLICATION_DOES_NOT_EXIST"]),
         },
@@ -101,7 +83,7 @@ class TablesView(APIView):
         database = DatabaseHandler().get_database(database_id)
         database.group.has_user(request.user, raise_error=True)
         tables = Table.objects.filter(database=database).prefetch_related("import_jobs")
-        serializer = TableWithImportJobsSerializer(tables, many=True)
+        serializer = TableSerializer(tables, many=True)
         return Response(serializer.data)
 
     @extend_schema(
@@ -124,16 +106,15 @@ class TablesView(APIView):
         ),
         request=TableCreateSerializer,
         responses={
-            200: TableWithImportJobsSerializer,
+            200: {
+                "type": "integer",
+                "description": "Id of the file import job for the new table.",
+                "example": 1,
+            },
             400: get_error_schema(
                 [
                     "ERROR_USER_NOT_IN_GROUP",
                     "ERROR_REQUEST_BODY_VALIDATION",
-                    "ERROR_INVALID_INITIAL_TABLE_DATA",
-                    "ERROR_INITIAL_TABLE_DATA_LIMIT_EXCEEDED",
-                    "ERROR_RESERVED_BASEROW_FIELD_NAME",
-                    "ERROR_INITIAL_TABLE_DATA_HAS_DUPLICATE_NAMES",
-                    "ERROR_INVALID_BASEROW_FIELD_NAME",
                     "ERROR_MAX_JOB_COUNT_EXCEEDED",
                 ]
             ),
@@ -145,28 +126,36 @@ class TablesView(APIView):
         {
             ApplicationDoesNotExist: ERROR_APPLICATION_DOES_NOT_EXIST,
             UserNotInGroup: ERROR_USER_NOT_IN_GROUP,
-            InvalidInitialTableData: ERROR_INVALID_INITIAL_TABLE_DATA,
-            InitialTableDataLimitExceeded: ERROR_INITIAL_TABLE_DATA_LIMIT_EXCEEDED,
-            MaxFieldLimitExceeded: ERROR_MAX_FIELD_COUNT_EXCEEDED,
-            MaxFieldNameLengthExceeded: ERROR_MAX_FIELD_NAME_LENGTH_EXCEEDED,
-            InitialTableDataDuplicateName: ERROR_INITIAL_TABLE_DATA_HAS_DUPLICATE_NAMES,
-            ReservedBaserowFieldNameException: ERROR_RESERVED_BASEROW_FIELD_NAME,
-            InvalidBaserowFieldName: ERROR_INVALID_BASEROW_FIELD_NAME,
             MaxJobCountExceeded: ERROR_MAX_JOB_COUNT_EXCEEDED,
         }
     )
     @validate_body(TableCreateSerializer)
     def post(self, request, data, database_id):
-        """Creates a new table in a database."""
+        """Creates a job to create a new table in a database."""
 
         database = DatabaseHandler().get_database(database_id)
 
-        table = action_type_registry.get_by_type(CreateTableActionType).do(
-            request.user, database, fill_example=True, **data
-        )
+        session = get_untrusted_client_session_id(request.user)
 
-        serializer = TableWithImportJobsSerializer(table)
-        return Response(serializer.data)
+        if not data["data"]:
+            file_import_job = TableHandler().create_minimal_table(
+                request.user,
+                database,
+                data["name"],
+                fill_example=True,
+                session=session,
+            )
+        else:
+            file_import_job = TableHandler().create_table(
+                request.user,
+                database,
+                data["name"],
+                data=data["data"],
+                first_row_header=data["first_row_header"],
+                session=session,
+            )
+
+        return Response(file_import_job.id)
 
 
 class TableView(APIView):
@@ -188,7 +177,7 @@ class TableView(APIView):
             "related database's group."
         ),
         responses={
-            200: TableWithImportJobsSerializer,
+            200: TableSerializer,
             400: get_error_schema(["ERROR_USER_NOT_IN_GROUP"]),
             404: get_error_schema(["ERROR_TABLE_DOES_NOT_EXIST"]),
         },
@@ -204,7 +193,7 @@ class TableView(APIView):
 
         table = TableHandler().get_table(table_id)
         table.database.group.has_user(request.user, raise_error=True)
-        serializer = TableWithImportJobsSerializer(table)
+        serializer = TableSerializer(table)
         return Response(serializer.data)
 
     @extend_schema(