+++ backend/tests/baserow/contrib/database/api/tables/test_table_views.py
@@ -1,6 +1,8 @@
 import pytest
+import json
 from unittest.mock import patch
 from django.db import connection
+from baserow.contrib.database.file_import.models import FileImportJob
 
 from rest_framework.status import (
     HTTP_200_OK,
@@ -10,10 +12,8 @@ from rest_framework.status import (
 )
 
 from django.shortcuts import reverse
-from django.conf import settings
-from django.test.utils import override_settings, CaptureQueriesContext
+from django.test.utils import CaptureQueriesContext
 
-from baserow.contrib.database.fields.models import TextField
 from baserow.contrib.database.table.models import Table
 
 
@@ -89,20 +89,6 @@ def test_create_table(api_client, data_fixture):
     database = data_fixture.create_database_application(user=user)
     database_2 = data_fixture.create_database_application()
 
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url, {"name": "Test 1"}, format="json", HTTP_AUTHORIZATION=f"JWT {token}"
-    )
-    assert response.status_code == HTTP_200_OK
-    json_response = response.json()
-
-    Table.objects.all().count() == 1
-    table = Table.objects.filter(database=database).first()
-
-    assert table.order == json_response["order"] == 1
-    assert table.name == json_response["name"]
-    assert table.id == json_response["id"]
-
     url = reverse("api:database:tables:list", kwargs={"database_id": database_2.id})
     response = api_client.post(
         url, {"name": "Test 1"}, format="json", HTTP_AUTHORIZATION=f"JWT {token}"
@@ -124,15 +110,34 @@ def test_create_table(api_client, data_fixture):
     assert response.status_code == HTTP_404_NOT_FOUND
     assert response.json()["error"] == "ERROR_APPLICATION_DOES_NOT_EXIST"
 
+    # Should create an example database
+    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
+    response = api_client.post(
+        url, {"name": "Test 1"}, format="json", HTTP_AUTHORIZATION=f"JWT {token}"
+    )
+    assert response.status_code == HTTP_200_OK
+    json_response = response.json()
 
-@pytest.mark.django_db
-@override_settings(FEATURE_FLAGS=[])
+    job = FileImportJob.objects.get(id=json_response)
+
+    assert job.table is not None
+    assert job.table.name == "Test 1"
+
+    model = job.table.get_model()
+
+    assert job.table.field_set.count() == 3
+    assert model.objects.count() == 2
+
+
+@pytest.mark.django_db(transaction=True)
 @patch("baserow.core.jobs.handler.run_async_job")
-def test_create_table_with_data(mock_run_async_job, api_client, data_fixture):
+def test_create_table_with_data(
+    mock_run_async_job, api_client, data_fixture, patch_filefield_storage
+):
     user, token = data_fixture.create_user_and_token()
     database = data_fixture.create_database_application(user=user)
-
     url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
+
     response = api_client.post(
         url,
         {"name": "Test 1", "data": []},
@@ -144,197 +149,18 @@ def test_create_table_with_data(mock_run_async_job, api_client, data_fixture):
     assert response_json["error"] == "ERROR_REQUEST_BODY_VALIDATION"
     assert response_json["detail"]["data"][0]["code"] == "min_length"
 
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {"name": "Test 1", "data": [[]]},
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INVALID_INITIAL_TABLE_DATA"
-
-    limit = settings.INITIAL_TABLE_DATA_LIMIT
-    settings.INITIAL_TABLE_DATA_LIMIT = 2
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {"name": "Test 1", "data": [[], [], []]},
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INITIAL_TABLE_DATA_LIMIT_EXCEEDED"
-    settings.INITIAL_TABLE_DATA_LIMIT = limit
-
-    field_limit = settings.MAX_FIELD_LIMIT
-    settings.MAX_FIELD_LIMIT = 2
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {"name": "Test 1", "data": [["fields"] * 3, ["rows"] * 3]},
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_MAX_FIELD_COUNT_EXCEEDED"
-    settings.MAX_FIELD_LIMIT = field_limit
-
-    too_long_field_name = "x" * 256
-    field_name_with_ok_length = "x" * 255
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 1",
-            "data": [
-                [too_long_field_name, "B", "C", "D"],
-                ["1-1", "1-2", "1-3", "1-4", "1-5"],
-                ["2-1", "2-2", "2-3"],
-                ["3-1", "3-2"],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_MAX_FIELD_NAME_LENGTH_EXCEEDED"
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 1",
-            "data": [
-                [field_name_with_ok_length, "B", "C", "D"],
-                ["1-1", "1-2", "1-3", "1-4", "1-5"],
-                ["2-1", "2-2", "2-3"],
-                ["3-1", "3-2"],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_200_OK
-
-    table = Table.objects.get(id=response_json["id"])
-
-    text_fields = TextField.objects.filter(table=table)
-    assert text_fields[0].name == field_name_with_ok_length
-    assert text_fields[1].name == "B"
-    assert text_fields[2].name == "C"
-    assert text_fields[3].name == "D"
-    assert text_fields[4].name == "Field 5"
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 1",
-            "data": [
-                ["A", "B", "C", "D"],
-                ["1-1", "1-2", "1-3", "1-4", "1-5"],
-                ["2-1", "2-2", "2-3"],
-                ["3-1", "3-2"],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_200_OK
-
-    table = Table.objects.get(id=response_json["id"])
-
-    text_fields = TextField.objects.filter(table=table)
-    assert text_fields[0].name == "A"
-    assert text_fields[1].name == "B"
-    assert text_fields[2].name == "C"
-    assert text_fields[3].name == "D"
-    assert text_fields[4].name == "Field 5"
-
-    model = table.get_model()
-    results = model.objects.all()
-
-    assert results.count() == 3
-
-    assert getattr(results[0], f"field_{text_fields[0].id}") == "1-1"
-    assert getattr(results[0], f"field_{text_fields[1].id}") == "1-2"
-    assert getattr(results[0], f"field_{text_fields[2].id}") == "1-3"
-    assert getattr(results[0], f"field_{text_fields[3].id}") == "1-4"
-    assert getattr(results[0], f"field_{text_fields[4].id}") == "1-5"
-
-    assert getattr(results[1], f"field_{text_fields[0].id}") == "2-1"
-    assert getattr(results[1], f"field_{text_fields[1].id}") == "2-2"
-    assert getattr(results[1], f"field_{text_fields[2].id}") == "2-3"
-    assert getattr(results[1], f"field_{text_fields[3].id}") == ""
-    assert getattr(results[1], f"field_{text_fields[4].id}") == ""
-
-    assert getattr(results[2], f"field_{text_fields[0].id}") == "3-1"
-    assert getattr(results[2], f"field_{text_fields[1].id}") == "3-2"
-    assert getattr(results[2], f"field_{text_fields[2].id}") == ""
-    assert getattr(results[2], f"field_{text_fields[3].id}") == ""
-    assert getattr(results[2], f"field_{text_fields[4].id}") == ""
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 2",
-            "data": [
-                ["1-1"],
-                ["2-1", "2-2", "2-3"],
-                ["3-1", "3-2"],
-            ],
-            "first_row_header": False,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_200_OK
-
-    table = Table.objects.get(id=response_json["id"])
-
-    text_fields = TextField.objects.filter(table=table)
-    assert text_fields[0].name == "Field 1"
-    assert text_fields[1].name == "Field 2"
-    assert text_fields[2].name == "Field 3"
-
-    model = table.get_model()
-    results = model.objects.all()
-
-    assert results.count() == 3
-
-    assert getattr(results[0], f"field_{text_fields[0].id}") == "1-1"
-    assert getattr(results[0], f"field_{text_fields[1].id}") == ""
-    assert getattr(results[0], f"field_{text_fields[2].id}") == ""
-
-    assert getattr(results[1], f"field_{text_fields[0].id}") == "2-1"
-    assert getattr(results[1], f"field_{text_fields[1].id}") == "2-2"
-    assert getattr(results[1], f"field_{text_fields[2].id}") == "2-3"
-
-    assert getattr(results[2], f"field_{text_fields[0].id}") == "3-1"
-    assert getattr(results[2], f"field_{text_fields[1].id}") == "3-2"
-
-    with override_settings(FEATURE_FLAGS=["async_import"]):
+    with patch_filefield_storage():
         response = api_client.post(
             url,
             {
-                "name": "Test 22",
+                "name": "Test 1",
                 "data": [
-                    ["1-1"],
+                    ["A", "B", "C", "D"],
+                    ["1-1", "1-2", "1-3", "1-4", "1-5"],
+                    ["2-1", "2-2", "2-3"],
+                    ["3-1", "3-2"],
                 ],
-                "first_row_header": False,
+                "first_row_header": True,
             },
             format="json",
             HTTP_AUTHORIZATION=f"JWT {token}",
@@ -342,122 +168,24 @@ def test_create_table_with_data(mock_run_async_job, api_client, data_fixture):
     response_json = response.json()
     assert response.status_code == HTTP_200_OK
 
-    assert "import_jobs" in response_json
-    assert len(response_json["import_jobs"]) == 1
+    mock_run_async_job.delay.assert_called_with(response_json)
 
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 2",
-            "data": [
-                [
-                    "TEst 1",
-                    "10.00",
-                    'Falsea"""',
-                    'a"a"a"a"a,',
-                    "a",
-                    "/w. r/awr",
-                ],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_200_OK
+    job = FileImportJob.objects.get(id=response_json)
 
-    table = Table.objects.get(id=response_json["id"])
-    text_fields = TextField.objects.filter(table=table)
-    assert text_fields[0].name == "TEst 1"
-    assert text_fields[1].name == "10.00"
-    assert text_fields[2].name == 'Falsea"""'
-    assert text_fields[3].name == 'a"a"a"a"a,'
-    assert text_fields[4].name == "a"
-    assert text_fields[5].name == "/w. r/awr"
+    assert job.table is None
+    assert job.name == "Test 1"
+    assert job.first_row_header
+    assert job.database == database
 
-    model = table.get_model()
-    results = model.objects.all()
-    assert results.count() == 0
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 4",
-            "data": [
-                [
-                    "id",
-                ],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_RESERVED_BASEROW_FIELD_NAME"
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 4",
-            "data": [
-                [
-                    "test",
-                    "test",
-                ],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INITIAL_TABLE_DATA_HAS_DUPLICATE_NAMES"
-    assert "unique" in response_json["detail"]
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 4",
-            "data": [
-                [
-                    " ",
-                ],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    response_json = response.json()
-    assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INVALID_BASEROW_FIELD_NAME"
-    assert "blank" in response_json["detail"]
-
-    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
-    response = api_client.post(
-        url,
-        {
-            "name": "Test 4",
-            "data": [
-                [
-                    " test 1",
-                    "  test 2",
-                ],
-            ],
-            "first_row_header": True,
-        },
-        format="json",
-        HTTP_AUTHORIZATION=f"JWT {token}",
-    )
-    assert response.status_code == HTTP_200_OK
+    with patch_filefield_storage():
+        with job.data_file.open("r") as fin:
+            data = json.load(fin)
+            assert data == [
+                ["A", "B", "C", "D"],
+                ["1-1", "1-2", "1-3", "1-4", "1-5"],
+                ["2-1", "2-2", "2-3"],
+                ["3-1", "3-2"],
+            ]
 
 
 @pytest.mark.django_db