+++ web-frontend/modules/database/components/table/TablePasteImporter.vue
@@ -58,7 +58,7 @@ export default {
   data() {
     return {
       values: {
-        data: '',
+        getData: null,
         firstRowHeader: true,
       },
       content: '',
@@ -68,33 +68,32 @@ export default {
   },
   validations: {
     values: {
-      data: { required },
+      getData: { required },
     },
     content: { required },
   },
   methods: {
     changed(content) {
+      this.$emit('changed')
       this.content = content
       this.reload()
     },
     reload() {
       if (this.content === '') {
-        this.values.data = ''
+        this.values.getData = null
         this.error = ''
         this.preview = {}
-        this.$emit('input', this.value)
         return
       }
 
       const limit = this.$env.INITIAL_TABLE_DATA_LIMIT
       const count = this.content.split(/\r\n|\r|\n/).length
       if (limit !== null && count > limit) {
-        this.values.data = ''
+        this.values.getData = null
         this.error = this.$t('tablePasteImporter.limitError', {
           limit,
         })
         this.preview = {}
-        this.$emit('input', this.value)
         return
       }
 
@@ -104,22 +103,29 @@ export default {
           // If parsed successfully and it is not empty then the initial data can be
           // prepared for creating the table. We store the data stringified because it
           // doesn't need to be reactive.
+          const rows = [...data.data]
           const dataWithHeader = this.ensureHeaderExistsAndIsValid(
             data.data,
             this.values.firstRowHeader
           )
-          this.values.data = JSON.stringify(dataWithHeader)
+
+          this.values.getData = () => {
+            return new Promise((resolve) => {
+              if (this.values.firstRowHeader) {
+                rows[0] = dataWithHeader[0]
+              }
+              resolve(rows)
+            })
+          }
           this.error = ''
           this.preview = this.getPreview(dataWithHeader)
-          this.$emit('input', this.value)
         },
         error(error) {
           // Papa parse has resulted in an error which we need to display to the user.
           // All previously loaded data will be removed.
-          this.values.data = ''
+          this.values.getData = null
           this.error = error.errors[0].message
           this.preview = {}
-          this.$emit('input', this.value)
         },
       })
     },