+++ changelog.md
@@ -30,6 +30,7 @@ For example:
 * Add startup check ensuring BASEROW_PUBLIC_URL and related variables are correct. [#1041](https://gitlab.com/bramw/baserow/-/issues/1041)
 * Made it possible to extend the register page.
 * Made it possible to extend the app layout.
+* Allow to import more than 15Mb. [949](ttps://gitlab.com/bramw/baserow/-/issues/949)
 
 ### Bug Fixes